refactor: share oauth buttons
